requirements.txt
streamlit
pdfplumber
python-docx
google-generativeai
scikit-learn
python-dotenv
streamlit-lottiettie